consolidating all_test_commands.py
