+++ craftassist/test/test_conditions.py
@@ -5,7 +5,7 @@ import unittest
 
 from copy import deepcopy
 from base_craftassist_test_case import BaseCraftassistTestCase
-from all_test_commands import *
+from base_agent.test.all_test_commands import *
 from fake_mobs import LoopMob, make_mob_opts
 from base_agent.base_util import TICKS_PER_SEC
