+++ craftassist/test/test_get_memory.py
@@ -4,7 +4,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import unittest
 import craftassist.agent.shapes as shapes
 from base_craftassist_test_case import BaseCraftassistTestCase
-from all_test_commands import *
+from base_agent.test.all_test_commands import *
 
 
 class GetMemoryTestCase(BaseCraftassistTestCase):
