+++ craftassist/test/test_put_memory.py
@@ -6,7 +6,7 @@ import unittest
 
 import craftassist.agent.shapes as shapes
 from base_craftassist_test_case import BaseCraftassistTestCase
-from all_test_commands import *
+from base_agent.test.all_test_commands import *
 
 
 class PutMemoryTestCase(BaseCraftassistTestCase):
