+++ locobot/test/test_interpreter_mock.py
@@ -13,7 +13,7 @@ import numpy as np
 
 from base_fakeagent_test_case import BaseFakeAgentTestCase
 import rotation
-from all_test_commands import MOVE_COMMANDS, GROUND_TRUTH_PARSES
+from base_agent.test.all_test_commands import MOVE_COMMANDS, GROUND_TRUTH_PARSES
 from test_utils import assert_turn_degree
 
 CUBE1 = (9, 0, 4)
